model fix

+++ database/migrations/2019_09_01_095418_create_chapters_table.php
@@ -18,7 +18,7 @@ class CreateChapterTable extends Migration
             $table->string('number')->unique();
             $table->string('name');
             $table->text('description')->nullable();
-            $table->bigInteger('parent_id')->nullable();
+            $table->unsignedInteger('parent_id')->nullable();
             $table->timestamps();
         });
     }
